refactor(core): Move Registy and Heartbeat logic
