--- agent/app/utils/registry/__init__.py
@@ -1,4 +1,0 @@
-"""
-Registry utilities for agent registration.
-"""
-
