+++ core/agent/app/utils/registry/heartbeat_logic.py
