+++ core/agent/app/utils/registry/onchain_registration.py
