+++ core/agent/app/utils/registry_client.py
